requirements2.txt
numpy
PyPDF2
faiss-cpu
sentence-transformers
google-generativeai